requirements.txt
jinja2
PyExifTool==0.1.1
